requirements.txt
amqp==5.2.0
APScheduler==3.10.4
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.2
billiard==4.2.0
bs4==0.0.1
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
convert_numbers==0.4
cron-descriptor==1.4.0
deprecation==2.1.0
distlib==0.3.8
Django==5.0
django-celery-beat
django-celery-results==2.5.1
django-crispy-forms==2.1
django-crontab==0.7.1
django-email-verification==0.3.3
django-jalali==6.0.1
django-jalali-date==1.1.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
filelock==3.13.1
h11==0.14.0
idna==3.6
ippanel==2.0.7
jdatetime==4.1.1
kombu==5.3.4
numpy==1.26.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.3
pause==0.3
Pillow==10.1.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
pycparser==2.21
PyJWT==2.8.0
pyotp==2.9.0
PySocks==1.7.1
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
redis-cli==1.0.1
requests==2.31.0
schedule==1.2.1
scheduler==0.8.4
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
trio==0.23.2
trio-websocket==0.11.1
typeguard==4.1.5
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
vine==5.1.0
virtualenv==20.25.0
wcwidth==0.2.12
wsproto==1.2.0
